refactor(@angular/cli): move MCP list projects tool to separate file
Move the list projects MCP tool for the Angular CLI's MCP server to a
separate file within the tools subdirectory. This provides a more
consistent structure that matches the other tools in the server.
